Move resolve/apply_resolved_bindings impl to descriptor_patching.cpp
The fast cache-hit patching logic (resolve_bindings, apply_resolved_bindings,
and the rt_binding_key helper) now lives in its own translation unit
tt_metal/impl/program/descriptor_patching.cpp, keeping program_descriptors.cpp
free of experimental-feature code. Declarations remain in
tt_metal/api/tt-metalium/experimental/descriptor_patching.hpp.
